+++ storage/cms/cache/01/71/avantage.htm.php
@@ -1,0 +1,5 @@
+<?php 
+class Cms5fd9ae188c56f158151315_4b9cbc97097adfe6cd3bfd4ca2a971daClass extends Cms\Classes\PartialCode
+{
+
+}