+++ storage/cms/cache/25/e5/zerospam.htm.php
@@ -1,0 +1,8 @@
+<?php 
+class Cms5fd9af37e0ef7684088576_903f53886b22c9496e053bc1a2200e6eClass extends Cms\Classes\PageCode
+{
+public function onEnd()
+{
+$this['image_url'] = asset('/themes/techno24/assets/images/0spam-com.jpg');
+}
+}