+++ storage/cms/cache/56/43/subscribe.htm.php
@@ -1,5 +1,0 @@
-<?php 
-class Cms5f1700ee32487402073521_72d4f147598becb05bf094f531b227ebClass extends Cms\Classes\PartialCode
-{
-
-}