+++ storage/cms/cache/64/ed/default.htm.php
@@ -1,0 +1,5 @@
+<?php 
+class Cms5fd9ad3b487c7151865528_9e0db9dcb8ee6c1ea1305531a845fe7dClass extends Cms\Classes\PartialCode
+{
+
+}