+++ storage/cms/cache/7a/e1/default.htm.php
@@ -1,0 +1,5 @@
+<?php 
+class Cms5fd9acaaba08b071525607_7ea831436beb928fd3fbf0c76ed88630Class extends Cms\Classes\PartialCode
+{
+
+}