+++ storage/cms/cache/82/1f/avantage.htm.php
@@ -1,0 +1,5 @@
+<?php 
+class Cms5fd9ad3b45628432737044_880d37f63a36c8b195484e9a59a3607eClass extends Cms\Classes\PartialCode
+{
+
+}