+++ storage/cms/cache/89/6d/default.htm.php
@@ -1,6 +1,6 @@
 <?php 
 use RainLab\Translate\Classes\Translator;
-class Cms5f1700e8559b0825779956_1eb0aaf94b76e2ee3b9337e0a62f845dClass extends Cms\Classes\LayoutCode
+class Cms5fd9abc0ea1f2549643757_727f48187498fd556dd4c42005e18dc2Class extends Cms\Classes\LayoutCode
 {
 public function onInit()
 {