+++ storage/cms/cache/95/73/default.htm.php
@@ -1,5 +1,0 @@
-<?php 
-class Cms5f1700ee1d1af001136677_02b79c584bfe6d76cc8514c11d99b0c4Class extends Cms\Classes\PartialCode
-{
-
-}