+++ storage/cms/cache/9a/83/hosting.htm.php
@@ -1,0 +1,5 @@
+<?php 
+class Cms5fd9ae86939f4629863300_0773f126094cbb27424fb5b9dcab3c4cClass extends Cms\Classes\PageCode
+{
+
+}