+++ storage/cms/cache/c3/26/footer.htm.php
@@ -1,5 +1,5 @@
 <?php 
-class Cms5f1700e93bdc5276012059_709ea0874db2464263afb9d8a5a56814Class extends Cms\Classes\PartialCode
+class Cms5fd9abf09a824570770944_10566c3b3d902f2e6deea1e1f73bb160Class extends Cms\Classes\PartialCode
 {
 
 }