+++ storage/cms/cache/dc/4c/questions.htm.php
@@ -1,0 +1,5 @@
+<?php 
+class Cms5fd9aedfd75e9302931948_5654b8113339b27dc4bca0cd6ee5229eClass extends Cms\Classes\PartialCode
+{
+
+}