+++ storage/cms/cache/fb/ed/navbar.htm.php
@@ -1,5 +1,5 @@
 <?php 
-class Cms5f1700e92b0b8364975019_0a1d89803b5cf2dac7d2dddb7dd43e90Class extends Cms\Classes\PartialCode
+class Cms5fd9abf08598d083673646_6a206443a5f7410d7dcff6db49f1cc2fClass extends Cms\Classes\PartialCode
 {
 
 }