+++ storage/cms/twig/00/0017215c4644e98d52ff84997cdc78bb0851094592dfc1d9d7465204b6f22402.php
@@ -1,0 +1,108 @@
+<?php
+
+use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Extension\SandboxExtension;
+use Twig\Markup;
+use Twig\Sandbox\SecurityError;
+use Twig\Sandbox\SecurityNotAllowedTagError;
+use Twig\Sandbox\SecurityNotAllowedFilterError;
+use Twig\Sandbox\SecurityNotAllowedFunctionError;
+use Twig\Source;
+use Twig\Template;
+
+/* /Users/alexandrefoisy/web sites/techno24/themes/techno24/partials/zerospam/intro.htm */
+class __TwigTemplate_8195af824b9135e1624a7179467cff6d5e2f8445e8a6cdd1d43b835ffff58558 extends \Twig\Template
+{
+    private $source;
+    private $macros = [];
+
+    public function __construct(Environment $env)
+    {
+        parent::__construct($env);
+
+        $this->source = $this->getSourceContext();
+
+        $this->parent = false;
+
+        $this->blocks = [
+        ];
+        $this->sandbox = $this->env->getExtension('\Twig\Extension\SandboxExtension');
+        $tags = array("content" => 12);
+        $filters = array("theme" => 5, "_" => 7);
+        $functions = array();
+
+        try {
+            $this->sandbox->checkSecurity(
+                ['content'],
+                ['theme', '_'],
+                []
+            );
+        } catch (SecurityError $e) {
+            $e->setSourceContext($this->source);
+
+            if ($e instanceof SecurityNotAllowedTagError && isset($tags[$e->getTagName()])) {
+                $e->setTemplateLine($tags[$e->getTagName()]);
+            } elseif ($e instanceof SecurityNotAllowedFilterError && isset($filters[$e->getFilterName()])) {
+                $e->setTemplateLine($filters[$e->getFilterName()]);
+            } elseif ($e instanceof SecurityNotAllowedFunctionError && isset($functions[$e->getFunctionName()])) {
+                $e->setTemplateLine($functions[$e->getFunctionName()]);
+            }
+
+            throw $e;
+        }
+
+    }
+
+    protected function doDisplay(array $context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        // line 1
+        echo "<div class=\"row my-5\">
+    <div class=\"col-xs-12 col-sm-12 col-lg-6 pb-3\">
+        <div class=\"secondary-shadow\">
+            <img class=\"img-fluid mx-auto d-block wow fadeInLeft\"
+                 src=\"";
+        // line 5
+        echo $this->extensions['Cms\Twig\Extension']->themeFilter("assets/images/0spam-com.jpg");
+        echo "\"
+                 data-holder-rendered=\"true\"
+                alt=\"";
+        // line 7
+        echo call_user_func_array($this->env->getFilter('_')->getCallable(), ["Capture écran de l'antivirus Bitdefender."]);
+        echo "\">
+        </div>
+    </div>
+    <div class=\"col\"></div>
+    <div class=\"col-sm-12 col-lg-5 wow fadeInUp\">
+        ";
+        // line 12
+        $context['__cms_content_params'] = [];
+        echo $this->env->getExtension('Cms\Twig\Extension')->contentFunction("produits/c-zerospam"        , $context['__cms_content_params']        );
+        unset($context['__cms_content_params']);
+        // line 13
+        echo "    </div>
+</div>";
+    }
+
+    public function getTemplateName()
+    {
+        return "/Users/alexandrefoisy/web sites/techno24/themes/techno24/partials/zerospam/intro.htm";
+    }
+
+    public function isTraitable()
+    {
+        return false;
+    }
+
+    public function getDebugInfo()
+    {
+        return array (  85 => 13,  81 => 12,  73 => 7,  68 => 5,  62 => 1,);
+    }
+
+    public function getSourceContext()
+    {
+        return new Source("", "/Users/alexandrefoisy/web sites/techno24/themes/techno24/partials/zerospam/intro.htm", "");
+    }
+}